% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethoscope_io.R
\name{buildEthoscopeQuery}
\alias{buildEthoscopeQuery}
\title{Build a query for loading ethoscope dat; using the date of experiments and devices name to retreive result files.}
\usage{
buildEthoscopeQuery(result_dir, query = NULL, use_cached = FALSE)
}
\arguments{
\item{result_dir}{The location of the result directory (i.e. the folder containing all the data).}

\item{query}{An optional query formatted as a dataframe (see details).}

\item{use_cached}{whether cache files should be used}
}
\value{
The query extended with the requested paths. When \code{query} is not specified, the function returns a table with all available files.
}
\description{
This function is designed to list and select experimental files. 
In general, end-users will want to retrieve  path to their experimental files
according to the date and ID of the video monitor without having to understand the underlying directory structure.
}
\details{
The optional argument \code{query} is expected to be a table where every row maps an experiment.
In many respects, it is similar to the \code{what} argument in \code{\link{loadEthoscopeData}}.
The only difference is that it does not have a \code{path} column. Instead, it must contain two columns:
\itemize{
 \item{\code{date} }{The date and time when the experiment started formatted either as `yyyy-mm-dd' or `yyyy-mm-dd_hh:mm:ss'.
 In the former case, there may be several matching experiments to a single time (starting the same day).
 When this happens, \emph{only the last} is returned, and a warning message is displayed.}
 \item{\code{machine_name} }{The name of the machine that acquired the data.}
}
The result is meant to be used directly, as the \code{what} argument, by \code{\link{loadEthoscopeData}} (see examples).
}
\note{
The ethoscope platform store data in a hard-coded directory structure \code{/root_dir/machine_id/machine_name/datetime/file.db}, where:
\itemize{
 \item{\code{machine_id} }{Is, in principle, a universally unique identifier of the acquisition device.}
 \item{\code{machine_name}, }{a human friendly name for acquisition device. In practice, this is expected to be unique within laboratory.}
 \item{\code{datetime}, }{the date and time of the start of the experiment}
}
}
\examples{
\dontrun{
# This is where I store the data on my computer
MY_DATA_DIR <- "/data/ethoscope_results/"

query <- data.table(date="2015-06-02",
                    machine_name=c("GGSM-001","GGSM-003"),
                    region_id = rep(1:10,each=2))
print(query)
map <- buildEthoscopeQuery(MY_DATA_DIR, query)
dt <- loadEthoscopeData(map)
}
}
\seealso{
\code{\link{cacheEthoscopeData}} to build a cached data directory.
}

