% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethoscope_io.R
\name{cacheEthoscopeData}
\alias{cacheEthoscopeData}
\title{Caches incrementally db  files to R native files}
\usage{
cacheEthoscopeData(result_dir, cached_dir, dry_run = F)
}
\arguments{
\item{result_dir}{The location of the result directory (i.e. the folder containing all the data).}

\item{cached_dir}{The location of the directory where data should be saved.}

\item{dry_run}{The location of the directory where data should be saved.}
}
\description{
This function is meant to be run by heavy users who want to speed up reading ethoscope file.
It will essentially preload all data in a result directory into a cached directory with the same structure.
.rdb and .idx fils are generated instead of .db file.
In practice, this function will be run every day to fetch and cache new data.
}
