% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{checkColumns}
\alias{checkColumns}
\title{Checks if the expected columns are all in a given character vector. Through error if not}
\usage{
checkColumns(expected_colnames, cols)
}
\arguments{
\item{expected_colnames}{the colnmaes that should be in a dt}

\item{cols}{the actual column names}
}
\description{
Checks if the expected columns are all in a given character vector. Through error if not
}
