% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sleep.R
\name{curateSparseRoiData}
\alias{curateSparseRoiData}
\title{remove data points when the time serie is too sparse}
\usage{
curateSparseRoiData(data, window = 60, min_points = 20)
}
\description{
remove data points when the time serie is too sparse
}

