% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dam_data}
\alias{dam_data}
\title{A simple rethomics data imported form a Trikinetics DAM2 file.}
\format{a \code{data.table} where each row is a single observation of one unique animal at one unique time}
\source{
Gilestro lab, Imperial college, London
}
\usage{
dam_data
}
\description{
A 32 animal, 5 days DAM2, single monitor data set formatted in a 
conventinal rethomics data.table (i.e. has \code{experiment_id}, \code{region_id} and \code{t} columns). 
A dummy variable, \code{condition}, was also added for illustration purposes.
}
\author{
Anne Petzold, 2015-07-02
}
