% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time.R
\name{days}
\alias{days}
\title{Trivially converts days to seconds}
\usage{
days(x)
}
\arguments{
\item{x}{number of days}
}
\value{
the corresponding number of seconds
}
\description{
Trivially converts days to seconds
}
\seealso{
\code{\link{hours}} \code{\link{mins}}
}

