% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{ethogramPlot}
\alias{ethogramPlot}
\title{TODO}
\usage{
ethogramPlot(data, var, condition, summary_time_window = 30 * 60,
  normalise_var_per_id = TRUE, error_bar = NULL)
}
\description{
TODO
TODO...... . .............. ... . .. . ......
}

