% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{fetchDAMData}
\alias{fetchDAMData}
\title{Retrieves DAM2 data from daily saved files}
\usage{
fetchDAMData(result_dir, query, reference_hour = 9, tz = "BST",
  FUN = NULL, ...)
}
\arguments{
\item{result_dir}{the root directory where all daily data are saved}

\item{query}{a formatted query used to request data (see detail).}

\item{reference_hour}{the hour, in the day, to use as t_0 reference. This should be expressed on Greenwich Meridian Time.}

\item{tz}{the time zone on which the DAM2 data was saved (e.g. BSM -> British Summer Time)}

\item{FUN}{an optional function to transform the data from each `region' (i.e. a data.table) immediately after is has been loaded.}

\item{...}{extra arguments to be passed to \code{FUN}}
}
\value{
A data.table where every row is an individual measurement. That is an activity at a unique time (\code{t}) in a
unique channel (\code{region_id}), and from a unique result date/experiment (\code{experiment_id}).
The time is expressed in seconds. For each different combination of \code{start_date} and \code{machine_id} in the query, an
individual \code{experiment_id} is generated.
}
\description{
Uses a query mechanism to get data from a DAM2 array.
This is useful when data has been saved, by day, in individual files for each monitor.
}
\details{
\code{query} must be a data.table.
Conceptually, each row of the query describes the conditions in one channel (when \code{region_id} is specified), or in each monitor (when it is not).
It should have  the following columns:
\itemize{
 \item{\code{machine_id} }{ the name of the machine used (e.g. `M002').}
 \item{\code{start_date} }{ the first day of the requested experiment (e.g. `2014-12-28').}
 \item{\code{stop_date} }{ the last day of the requested experiment (e.g. `2014-12-30').}
 \item{\code{region_id} }{ the channel (between 1 and 32) in what the animal was in (e.g. `20'). This is an optional column. If not provided, all 32 channels are loaded with the same conditions.}
 \item{\code{...} }{ arbitrary columns to associate conditions/treatments/genotypes/... to the previous columns}
}
}
\note{
the daily data should be saved in a hard-coded directory structure \code{root_dir/yyyy/mm/mmdd/mmddMxyz.txt}, where:
\itemize{
 \item{\code{yyyy} }{Is the year (e.g. 2014)}
 \item{\code{mm} and \code{dd}, }{the formatted month and day, respectively (e.g. mm=12 and dd=28).}
 \item{\code{xyz}, }{the number of the monitor (e.g 003)}
}
}
\seealso{
\code{\link{queryDAMData}} to load data from a regular DAM2 file
}

