% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{getSampleDataPath}
\alias{getSampleDataPath}
\title{Get the absolute path to a sample file.}
\usage{
getSampleDataPath(path = NULL, get_sample_dir = FALSE)
}
\arguments{
\item{path}{The relative path and name  of the samples to be loaded. When \code{path} is \code{NULL}, 
the function returns the list of all available samples.}

\item{get_sample_dir}{whether the function return the root directory of the sample data, nistead of sample files.}
}
\description{
This function is only for testing (and trying) purposes.
It provides a way to access raw data  (e.g. db files and dam text files) contained within this package.
}
\seealso{
\code{\link{loadEthoscopeData}} to read raw experimental data.
}
