% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggetho.R
\name{ggetho}
\alias{ggetho}
\title{Prepare a ggplot object to represent behavioural data}
\usage{
ggetho(variable, data, mapping, summary_FUN = mean,
  summary_time_window = mins(30), time_wrap = NULL, time_offset = NULL,
  ...)
}
\arguments{
\item{variable}{Variable of interest.}

\item{data}{\code{data.table} containing the data. It must have a column with the same name as \code{variable}.}

\item{mapping}{Default list of aesthetic mappings to use for plot.}

\item{summary_FUN}{Method (function) used to summarise \code{variable} over time (typically, the mean).}

\item{summary_time_window}{Width (in seconds) of the time window to compute a summary on.}

\item{time_wrap}{Time (in seconds) used to wrap the data (see details).}

\item{time_offset}{Time offset (i.e. phase, in seconds) when using \code{time_wrap}.}

\item{...}{Additional arguments to be passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
}
\value{
A initial plot object that can be further edited.
}
\description{
This function summarises a variable of interest
in order to subsequently represent it over time (using a panel of plotting functions).
}
\details{
\code{time_wrap} is typically used to express time relatively to the start of the the day.
In other words, it can help be used to pull all days together in one representative day.
In this case, \code{time_wrap=hours(24)}.
Instead of representing data from the start of the day, it can be done from any offset, using \code{time_offset}.
For instance,  \code{time_offset= hours(12)} puts the circadian reference (ZT0) in the middle of the plot.
}
\examples{
# we start by making a to dataset with 20 animals
query<- data.table(experiment_id="toy_experiment",
                   region_id=1:20, 
                   condition=c("A","B"))
dt <- toyActivityData(query,3)
# We build a plot object
pl <-  ggetho(asleep,     # the variable of interest
         dt,              # our original data
         aes(
             x=t,         # we have the time on the x axis
             y=asleep     # and asleep on the y axis
             )     
         ) 
# Sometimes, the variable of interest in not on the y.
When we do not provide a y axis,
ggetho will make a ID fo each animal and display them on separate rows
pl <-  ggetho(asleep,dt, aes(x=t)) 

}
\seealso{
\link{stat_pop_etho} To show trend by aggregating individuals over time.
}
\author{
Quentin Geissmann (\email{qgeissmann@gmail.com})
}
