% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interpol.R
\name{interpolateROIData}
\alias{interpolateROIData}
\title{Resample ROI data to a regular multivariate time series.}
\usage{
interpolateROIData(data, fs)
}
\arguments{
\item{fs}{the desired resampling frequency, in Hz.}

\item{d}{a datatable containing a time column (\code{d$t}) and one or several columns for recorded variables.
Typicaly, this would be generated by  \code{link{loadROIsFromFile}}.}
}
\value{
A new data.table with the same columns as \code{d}.
}
\description{
This function computes average (arithmetic mean) of all recorded variables at regular intervals by windowing the time series.
Then, it performs linear interpolation in order to infer value of missing points.
}
\note{
The raw data is not guaranted to be a rugular time series in so far as
1. the exact time stamp of every raw data point depends on the performance of the acquisition device,
and 2. there could be missing data point when an animal is ocluded or not detected
}
\examples{
\dontrun{
FILE <- "result.db"
# Load the three first ROIs
ldfs <- loadROIsFromFile(FILE, rois=1:3)
###### Simple example resampling the first ROI at 0.1Hz (i.e. every 10s):
d <- ldfs[roi_id == 2,]
new_d <- interpolateROIData(d, fs=0.1)
print(new_d)
}
}
\seealso{
\code{\link{loadROIsFromFile}} in order to load ROI data.
}

