% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dam2-data.R
\name{loadDAM2Data}
\alias{loadDAM2Data}
\title{Retrieves DAM2 data from continuous files}
\usage{
loadDAM2Data(query, FUN = NULL, ...)
}
\arguments{
\item{query}{Formatted query used to request data (see detail).}

\item{FUN}{Optional function to transform the data from each "region" (i.e. a \code{data.table}) immediately after is has been loaded.}

\item{...}{Extra arguments to be passed to \code{FUN}}
}
\value{
A \code{data.table} where every row is an individual measurement.
That is an activity at a unique time (\code{t}, in second) in a unique channel (\code{region_id}),
and from a unique date/experiment (\code{experiment_id}).
For each different combination of \code{start_date} and \code{file} in the query, an
individual \code{experiment_id} is generated.
}
\description{
Uses a query mechanism to get data from a DAM2 array.
This is useful when using the default behaviour of Trikinetics software
where data is simply appended to a single long file per monitor.
}
\details{
\code{query} must be a \code{data.table}.
Conceptually, each row of the query describes the conditions in one channel (when \code{region_id} is specified),
or in each monitor (when it is not).
It should have  the following columns:
\itemize{
\item \code{path} the location of your data file (e.g. \code{"C:/User/me/Desktop/Monitor3.txt"})
\item \code{start_date} the first day of the requested experiment (e.g. \code{"2014-12-28_10-00-00"}).
Importantly, the time component (i.e. \code{"10-00-00"}in this example) is used as a reference time.
Here, ZT0 will be \code{10:00:00} in the DAM file.
\item \code{stop_date}  the last day of the requested experiment (e.g. \code{"2014-12-30"}).
\item \code{region_id`` the channel (between 1 and 32) in what the animal was in (e.g.}20`).
This is an optional column. If not provided, all 32 channels are loaded with the same conditions.
\item \code{...} arbitrary columns to associate conditions/treatments/genotypes/... to the previous columns
}
}
\examples{
\dontrun{
# This is a simple, single sample dam2 file:
sample_file <- getSampleDataPath('misc/DAMfile.dam')
query = data.table(path=sample_file,
                   # note the time (10:00) is added as reference time
                 start_date="2015-07-02_10-00-00", 
                 stop_date="2015-07-07",
                 region_id=c(1:32),condition=rep(letters[1:2],each=16))
print(query)
dt <- loadDAM2Data(query)
ethogramPlot(activity,dt,condition) + scale_x_continuous(breaks=0:10/2)
dt <- loadDAM2Data(query,FUN= sleepDAMAnnotation)
ethogramPlot(asleep,dt,condition) + scale_x_continuous(breaks=0:10/2)
}
}
\seealso{
\itemize{
\item Tutorial for this function \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
\item What queries are \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
\item The structure of the resuting data \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
}
}
