% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/io.R
\name{loadDAMFile}
\alias{loadDAMFile}
\title{Read a text file formatted as DAM2 into a single data table.}
\usage{
loadDAMFile(FILE, channels = NULL, min_time = 0, max_time = Inf,
  interval = 60)
}
\arguments{
\item{FILE}{the name of the input file.}

\item{min_time}{exclude data before min_time (in seconds). This time is relative to the start of the experiement.}

\item{max_time}{exclude data after max_time (in seconds). This time is relative to the start of the experiement.}
}
\value{
If \code{rois} has only one element, a dataframe. Otherwise, a list of dataframes (one per ROI)
}
\description{
This function is used to load data from DAM2 devices as a data.table.
}
\note{
Analysis of many long (sevaral days) recording can use a large amount of RAM.
Therefore, it can sometimes be advantageaous to load an process ROIs one by one.
}
\examples{
\dontrun{
FILE <- "Monitor53.txt"
out <- loadDAMFile(FILE)
#histogram of x marginal distribution
hist(out[roi_id == 1, x], nclass=100)
}
\dontrun{
# More realistec example where we have experiemental conditions, and
we want to resample data at 1.0Hz.
# First, the conditions:
conditions <- cbind(roi_id=1:32, expand.grid(treatment=c(T,F), genotype=LETTERS[1:4]))
print(conditions)
}
}
\seealso{
\code{\link{loadMetaData}} To display global informations about the experiment.
}

