% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ethoscope-data.R
\name{loadEthoscopeMetaData}
\alias{loadEthoscopeMetaData}
\title{Retrieves  metadata from an ethoscope result file}
\usage{
loadEthoscopeMetaData(FILE)
}
\arguments{
\item{FILE}{the name of the input file.}
}
\value{
A list containing fields for metadata entries
}
\description{
This function is used to obtain metadata -- such as \emph{time and date} of the experiment ,
\emph{acquisition device}, \emph{version of the software} and such--
embedded in a result file generated by the ethoscope platform.
}
\seealso{
\itemize{
\item \link{loadEthoscopeData} to load the actual data
}
}
