% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{loadMetaData}
\alias{loadMetaData}
\title{Retreive metadata from a result file.}
\usage{
loadMetaData(FILE)
}
\arguments{
\item{FILE}{the name of the input file.}
}
\value{
A list containing fields for metadata entries
}
\description{
This function is used to obtain metadata -- such as `time and date of the experiment' , `acquisition device', `version of the software' and such--
embeded in a result file generated by PSV.
}
\examples{
\dontrun{
FILE <- "result.db"
out <- loadMetaData(FILE)
names(out)
}
}
\seealso{
\code{\link{loadPsvData}} to load raw data.
}

