% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethoscope_io.R
\name{loadOneROICached}
\alias{loadOneROICached}
\title{A function to load data from a .rdb file this is faster cause it is indexed and R binary data.
See cacheEthoscopeData in order to generate cached files}
\usage{
loadOneROICached(FILE, region_id, min_time = 0, max_time = Inf,
  reference_hour = NULL, columns = NULL)
}
\description{
A function to load data from a .rdb file this is faster cause it is indexed and R binary data.
See cacheEthoscopeData in order to generate cached files
}

