% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{makeBoutDt}
\alias{makeBoutDt}
\title{set.seed(1)
bout_length <- round(runif(100,1,100))
bout_val <- rep(c(T,F),length.out=length(bout_length))
x <- rep(bout_val,bout_length)
test_dt  <- data.table(x=x,
  t=1:length(x) + 50)
makeBoutDt(x,test_dt)}
\usage{
makeBoutDt(x, sub_data)
}
\description{
set.seed(1)
bout_length <- round(runif(100,1,100))
bout_val <- rep(c(T,F),length.out=length(bout_length))
x <- rep(bout_val,bout_length)
test_dt  <- data.table(x=x,
  t=1:length(x) + 50)
makeBoutDt(x,test_dt)
}

