% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makeLDAnnotation}
\alias{makeLDAnnotation}
\title{Put white and black bars under a plot to show Dark and Light phases.}
\usage{
makeLDAnnotation(pl, time_conversion_unit = days, period = hours(24),
  offset = 0, size = 0.02, colours = c("white", "black"))
}
\arguments{
\item{pl}{A \code{ggplot} object to be annotated.}

\item{time_conversion_unit}{The time conversion function used in \code{pl}.}

\item{period}{The period, in seconds}

\item{offset}{A number, between 0 and the \code{period} to shift the bars to the right}

\item{size}{The hight of the bar. It is expressed in percent of the graph height}

\item{colours}{A vector of two colours to be used for annotation (typically, white and black)}
}
\value{
A \code{ggplot} object that can be plotted directly, or modified.
}
\description{
Put white and black bars under a plot to show Dark and Light phases.
}
\examples{
data(sleep_sexual_dimorphism)
my_data <- sleep_sexual_dimorphism
# Fraction of animal asleep over time:
p <- overviewPlot(asleep,my_data,condition=sex)
p <- makeLDAnnotation(p)
print(p)
# if you want to use subjective days (DD):
p <- makeLDAnnotation(p, colours=c("grey", "black"))
print(p)
}
