% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethoscope_io.R
\name{makeMasterTable}
\alias{makeMasterTable}
\title{Generates a unified master table for any allowed `what` argument.
The resulting master table will always have the columns "path", "experiment_id", "region_id"}
\usage{
makeMasterTable(what)
}
\description{
Generates a unified master table for any allowed `what` argument.
The resulting master table will always have the columns "path", "experiment_id", "region_id"
}

