% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{maxVelocityClassifierLegacy}
\alias{maxVelocityClassifierLegacy}
\title{Motion classifier based on maximum velocity.}
\usage{
maxVelocityClassifierLegacy(data, velocity_threshold = 0.006)
}
\arguments{
\item{data}{the data.table containing behavioural features used for movement classification.}

\item{velocity_threshold}{velocity above which an animal is classified as `moving'.}
}
\value{
a data table with the columns \code{moving} (logical, TRUE iff. motion was detected) and \code{t_round} (the `rounded' time). There is one row per rounded time point.
}
\description{
Defines whether an animal is moving according to its subpixel velocity. 
It requires a variable named \code{xy_dist_log10x1000} in the .db file.
}
\seealso{
\code{\link{sleepAnnotation}} to apply this function to all subsequent time windows.
\code{\link{maxVelocityClassifierMasked}}, the emiprically corected FPS invariant version of this classifier.
}
