% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion-classifiers.R
\name{maxVelocityClassifierMasked}
\alias{maxVelocityClassifierMasked}
\alias{maxVelocityClassifierLegacy}
\alias{virtualBeamCrossClassif}
\title{Motion classifier for Ethocope}
\usage{
maxVelocityClassifierMasked(data, velocity_correction_coef = 0.003,
  masking_duration = 6)

maxVelocityClassifierLegacy(data, velocity_threshold = 0.006)

virtualBeamCrossClassif(data)
}
\arguments{
\item{data}{\code{data.table} containing behavioural features used for movement classification.
It must have the columns \code{xy_dist_log10x1000}(for computing subpixel velocity), \code{x}(beam cross), \code{t} and \code{t_round}
(time points belonging to the same scoring window share the same \code{t_round}).}

\item{velocity_correction_coef}{an empirical coefficient to correct velocity with respect
to variable framerate.}

\item{masking_duration}{the number of second when any movement is ignored (velocity is set to 0) after
a stimulus is delivered (aka interaction).}

\item{velocity_threshold}{Uncorrected velocity above which an animal is classified as `moving' (for the legacy version).}
}
\value{
a data table with the columns:
\itemize{
\item \code{moving} Logical, TRUE iff. motion was detected.
\item \code{t_round} The "rounded" time.
\item \code{beam_crosses} The number of beam crosses
(when the animal crosses x = 0.5 -- that is the midpoint of the region of interest) within the time window
\item \code{max_velocity} The maximal velocity within the time window.
}
}
\description{
Defines whether an animal is moving according to:
}
\details{
\itemize{
\item Validated and corrected subpixel velocity (\link{maxVelocityClassifierMasked}), the mostb rigorous.
\item Uncorrected subpixel velocity (\link{maxVelocityClassifierLegacy}).
\item Crossing a virtual beam in the middle of the region of interest (\link{virtualBeamCrossClassif}).
}

\link{maxVelocityClassifierMasked} is the default movement classification for real-time ethoscope experiments.
It is benchmarked against human-generated ground truth.

These functions are \emph{rarely used directly}, but rather used by \link{sleepAnnotation}.
}
\seealso{
\itemize{
\item Tutorial for sleep analysis with ethoscopes \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
\item \link{sleepAnnotation} that requieres a motion classifier
}
}
