% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{multiple_iterative_y_mazes}
\alias{multiple_iterative_y_mazes}
\title{Unprocessed ethoscope data of animals waling in a y maze.}
\format{a \code{data.table} where each row is a single observation of one unique animal at one unique time}
\source{
Gilestro lab, Imperial college, London
}
\usage{
multiple_iterative_y_mazes
}
\description{
An ,8 animal, single region_id data set formatted in a 
conventinal rethomics data.table (i.e. has \code{experiment_id}, \code{region_id} and \code{t} columns).
}
\author{
Diana Bicazan, 2015-08-05
}

