% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethoscope_io.R
\name{parseOneROI}
\alias{parseOneROI}
\title{we obtain data from one ROI and optionaly preanalyse it, by applying FUN.}
\usage{
parseOneROI(i, master_table, min_time, max_time, reference_hour, verbose, FUN,
  columns = NULL, ...)
}
\description{
we obtain data from one ROI and optionaly preanalyse it, by applying FUN.
}

