% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sleep_sexual_dimorphism}
\alias{sleep_sexual_dimorphism}
\title{Sample of ethoscope data showing difference between males and females.}
\format{a \code{data.table} where each row is a single observation of one unique animal at one unique time}
\source{
Gilestro lab, Imperial college, London
}
\usage{
sleep_sexual_dimorphism
}
\description{
A sleep-annotated 56 animal data set.Movement was defined for 10s time windows.
conventinal rethomics data.table (i.e. has \code{experiment_id}, \code{region_id} and \code{t} columns).
}
\author{
Quentin Geissmann, 2015-06-13
}
