% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleepAnnotation}
\alias{sleepAnnotation}
\title{Determine whether an animal is asleep}
\usage{
sleepAnnotation(data, time_window_length = 10, min_time_immobile = 60 * 5,
  motion_classifier_FUN = totalWlkdDistClassif, ...)
}
\arguments{
\item{data}{the data (i.e a data.table) from a \emph{single} ROI. It must contain, at least,
the columns `t`, `x` and `y`.}

\item{time_window_length}{The number of seconds tu use in the motion classifier}

\item{min_time_immobile}{the minimal duration (in s) after which an immobile an animal is scored as `asleep'.}
}
\value{
A data table similar to \code{data} with additionnal variables/annotations (i.e. `activity', `moving', `asleep').
}
\description{
This function uses a motion classifier to first define whether an animal is moving during a given time window.
Then, it defines sleep as contiguous immobility for a minimal duration.
}
\note{
The resulting data will only have one data point every \code{time_window_length} seconds.
}
\examples{
# We load samples from the package data
file <- loadSampleData("validation.db")
# We would like only ROI #2 from this file
map <- data.frame(path=file, roi_id=2)
dt <- loadPsvData(map)
sleep_dt <-  sleepAnnotation(dt)
# A more likely scenario, we load ROIs 5 to 10,
# apply sleep analysis in combination with loadPsvData.
# this means we apply the function to all rois just after they are being loaded.
map <- data.frame(path=file, roi_id=5:10)
dt <- loadPsvData(map,FUN=sleepAnalysis)
sleep_dt <- dt[,sleepAnnotation(.SD),by=key(dt)]
}
\seealso{
\code{\link{loadPsvData}} to load data and optionnaly apply analysis on the fly.
}

