% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleepDAMAnnotation}
\alias{sleepDAMAnnotation}
\title{Determines whether an animal is asleep using beam crossing activity}
\usage{
sleepDAMAnnotation(data, time_window_length = 60, min_time_immobile = 60 *
  5)
}
\arguments{
\item{data}{the data (i.e a data.table) from a \emph{single} region. It must contain, at least,
the columns \code{t}, \code{x} and \code{y}.}

\item{time_window_length}{The number of seconds to be used by the motion classifier. This corresponds to the sampling period of the output data.}

\item{min_time_immobile}{the minimal duration (in s) after which an immobile animal is scored as `asleep'.}
}
\value{
A data table similar to \code{data} with additional variables/annotations (i.e. `moving', `asleep').
}
\description{
Sleep as contiguous inactivity (absence of beam crossing) for a minimal duration.
}
\note{
The resulting data will only have one data point every \code{time_window_length} seconds.
}
\examples{
# Let us load some sample data
data(dam_data)
dam_data[,
           sleepDAMAnnotation(.SD),
           by=key(dam_data)]
}
\seealso{
\code{\link{queryDAMData}} to load data in and apply this function directly
}

