% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sleep.R
\name{totalWlkdDistClassif}
\alias{totalWlkdDistClassif}
\title{Define whether an animal is moving from behavioural data.}
\usage{
totalWlkdDistClassif(data, activity_threshold = 0.03)
}
\description{
Define whether an animal is moving from behavioural data.
}
\seealso{
\code{\link{sleepAnnotation}} to apply this function to all subsequent time windows.
}

