% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-data.R
\name{toyEthoscopeData}
\alias{toyEthoscopeData}
\title{Generate toy ethoscope data}
\usage{
toyEthoscopeData(query = NULL, ...)
}
\arguments{
\item{query}{Query (i.e. a \code{data.table} where every row defines an animal).
Typically queries have, at least, the columns \code{experiment_id} and \code{region_id}.
The default value (\code{NULL}), will generate data for a single animal.}

\item{...}{Additional arguments to be passed to \link{toyActivityData}}
}
\value{
A behavioural \code{data.table} with the query columns as key and
the behavioural variables \code{xy_dist_log10x1000}, \code{has_interacted} and \code{x}
}
\description{
This function generates random data that emulates data as it is read from ethoscope files.
This is designed \strong{exclusively to provide material for examples and tests} since it generates  "realistic" datasets of arbitrary length.
}
\examples{
query<- data.table(experiment_id="toy_experiment",
                   region_id=1:4, 
                   condition=c("A","B"))
print(query)
dt <- toyEthoscopeData(query,duration=days(3))
print(dt)
}
\author{
Quentin Geissmann (\email{qgeissmann@gmail.com})
}
