% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tube_monitor_validation}
\alias{tube_monitor_validation}
\title{A dataset to validate tracking algorithms.}
\format{a \code{data.table} where each row is a single observation of one unique animal at one unique time}
\source{
Gilestro lab, Imperial college, London
}
\usage{
tube_monitor_validation
}
\description{
A 20 animal  single monitor data set formatted in a 
conventinal rethomics data.table (i.e. has \code{experiment_id}, \code{region_id} and \code{t} columns). 
Only raw tracking values are present in dataset.
}
\author{
Quentin Geissmann, 2015-05-02
}
