% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sleep.R
\name{virtualBeamCrossClassif}
\alias{virtualBeamCrossClassif}
\title{Motion classifier based on beam crosses.}
\usage{
virtualBeamCrossClassif(data)
}
\arguments{
\item{data}{the data.table containing behavioural features used for movement classification.}
}
\description{
Defines whether an animal is moving. This is achieved by computing the number of crossed of
a "virtual beam" in the middle of its region (i.e. at x=0.5).
This emulate the type of data generated by DAM2.
}
\seealso{
\code{\link{maxVelocityClassifier}} to defince movement by maximum velocity, which is more accurate, instead.
}

