% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkColumns}
\alias{checkColumns}
\title{Checks if the expected columns are all in a given character vector. Through error if not}
\usage{
checkColumns(expected_colnames, cols)
}
\description{
Checks if the expected columns are all in a given character vector. Through error if not
}

