% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{days}
\alias{days}
\title{Trivially converts days to seconds}
\usage{
days(x)
}
\arguments{
\item{x}{number of seconds}
}
\description{
Trivially converts days to seconds
}
\seealso{
\code{\link{hours}} \code{\link{days}}
}

