% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ethogramPlot}
\alias{ethogramPlot}
\title{Displays the temporal and inter-individual average of a variable of interest.}
\usage{
ethogramPlot(y, data, condition = NULL, facet_var = NULL,
  summary_time_window = mins(30), normalise_var_per_id = FALSE,
  error_bar = NULL)
}
\arguments{
\item{y}{The variable of interest.}

\item{data}{The data.table containing the data. It must have a column with the same name as y.}

\item{condition}{An optional grouping factor to order rows.}

\item{summary_time_window}{the width (in seconds) of the time window used to draw each ``pixel''.}

\item{normalise_var_per_id}{whether each row is to be normalised (using \code{new_x = (x - mean(x))/sd(x)}).}

\item{error_bar}{what type of error bar should be used. It should be one of \code{NULL},`sem' or `sd'.}
}
\value{
A \code{ggplot} object that can be plotted directly, or modified.
}
\description{
This function produces a graph where the variable of interest and time are on the y and x axes, respectively.
It can be used to visualise temporal trends per groups of conditions.
The response variable, y,  is grouped by time windows of defined size.
}
\examples{
data(sleep_male_vs_females)
my_data <- sleep_male_vs_females
# Fraction of animal asleep over time:
p <- ethogramPlot(asleep,my_data)
# We would like to show that per group:
p <- ethogramPlot(asleep,my_data,condition=sex)
print(p)
# We can also put error bars:
p <- ethogramPlot(asleep,my_data,condition=sex,error_bar="sem")
print(p)
# we can also use a condition to split data per row (ggplot faceting):
p <- ethogramPlot(asleep,my_data,condition=sex,facet_var=experiment_id,error_bar="sem")
print(p)
# p is simply a ggplot object, so we can change things:
print(p + labs(title="MY own title"))
}
\seealso{
\code{\link{oveviewPlot}} to show per-individual patterns
}

