% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ethogramPlot}
\alias{ethogramPlot}
\title{Displays the temporal and inter-individual average of a variable of interest.}
\usage{
ethogramPlot(y, data, condition = NULL, facet_var = NULL,
  summary_time_window = mins(30), normalise_var_per_id = FALSE,
  error_bar = NULL, time_wrap = NULL, time_unit_conversion = days)
}
\arguments{
\item{y}{The variable of interest.}

\item{data}{The data.table containing the data. It must have a column with the same name as y.}

\item{condition}{An optional grouping factor to order rows.}

\item{facet_var}{An optional grouping factor to draw group in each row of a faceted plot}

\item{summary_time_window}{the width (in seconds) of the time window used to draw each ``pixel''.}

\item{normalise_var_per_id}{whether each row is to be normalised (using \code{new_x = (x - mean(x))/sd(x)}).}

\item{error_bar}{what type of error bar should be used see details.}

\item{time_wrap}{the time (in seconds) used to wrap the data (see details).}

\item{time_unit_conversion}{a function to convert time in the x axis. typically, \code{days}, \code{hours} or \code{mins}.}
}
\value{
A \code{ggplot} object that can be plotted directly, or modified.
}
\description{
This function produces a graph where the variable of interest and time are on the y and x axes, respectively.
It can be used to visualise temporal trends per groups of conditions.
The response variable, y,  is grouped by time windows of defined size.
}
\details{
\code{time_wrap} is typically used to express time relatively to the start of the the day.
In other words, it can help be used to pull all days together in one representative day. In this case, \code{time_wrap=hours(24)}`.

At the moment, four types of error bars (\code{error_bar}) are supported:
\itemize{
 \item{`sd' }{The standard error}
 \item{`sem' }{The standard error of the mean (\emph{i.e.} \eqn{\frac{sd}{\sqrt{n}}})}
 \item{`gauss_ci' }{The gaussian 95\% confidence interval (\emph{i.e.} \eqn{1.96 \cdot{} \frac{sd}{\sqrt{n}}})}
 \item{`boot_ci' }{A standard 95\% bootstrap resampling confidence interval.
                  This is done over 5000 replicates. This can be quite \emph{slow}, but is often more statistically sound.}
}
}
\examples{
data(sleep_sexual_dimorphism)
my_data <- sleep_sexual_dimorphism
# Fraction of animal asleep over time:
p <- ethogramPlot(asleep,my_data)
# We would like to show that per group:
p <- ethogramPlot(asleep,my_data,condition=sex)
print(p)
# We can also put error bars:
p <- ethogramPlot(asleep,my_data,condition=sex,error_bar="sem")
print(p)
# we can also use a condition to split data per row (ggplot faceting):
p <- ethogramPlot(asleep,my_data,condition=sex,facet_var=experiment_id,error_bar="sem")
print(p)
# p is simply a ggplot object, so we can change things:
print(p + labs(title="MY own title"))
# Let us play with several error bars:
p <- ethogramPlot(asleep,my_data,condition=sex,error_bar="sd")
p
p <- ethogramPlot(asleep,my_data,condition=sex,error_bar="sem")
p
p <- ethogramPlot(asleep,my_data,condition=sex,error_bar="gauss_ci")
p
# this one is a bit slow
p <- ethogramPlot(asleep,my_data,condition=sex,error_bar="boot_ci")
p
data(dam_data)
# Time, on the x axis, in hours via
p <- ethogramPlot(activity,
             dam_data,
             condition,
             error_bar = "sem",
             time_unit_conversion=hours # this is where you set time in hours
             )
p
# summarise/wrap data in one day
p <- ethogramPlot(activity,
             dam_data,
             condition,
             error_bar = "sem",
             time_wrap=days(1) # this argument does the job
             )
p
}
\seealso{
\code{\link{overviewPlot}} to show per-individual patterns
}

