% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{ethogramPlot}
\alias{ethogramPlot}
\title{Displays the temporal and inter-individual average of a variable of interest.}
\usage{
ethogramPlot(y, data, condition = NULL, summary_time_window = mins(30),
  normalise_var_per_id = FALSE, error_bar = NULL)
}
\arguments{
\item{y}{The variable of interest}

\item{data}{The data.table containing the data. It must have a column with the same name as y.}

\item{condition}{An optionnal grouping factor to order rows.}

\item{summary_time_window}{the width (in seconds) of the time window used to draw each pixel.}

\item{normalise_var_per_id}{whether each row is to be normalised (using new_x = {x - mean(x)}/sd(x)).}

\item{error_bar}{what type of error bar should be used. It should be one of \code{NULL},`sem' or `sd').}
}
\value{
A \code{ggplot} object that can be plotted directly, or modified
}
\description{
This function produces a graph where y is the variable of interest and x, the time.
It can be used to visualise temporal trends per groups of conditions.
The response variable, y,  is grouped by time windows of defined size.
}
\examples{
# Load sample data
data(sleep_validation)
my_data <- sleep_validation[,sleepAnnotation(.SD),by=key(sleep_validation)]
# No condition
p <- ethogamPlot(activity,my_data)
print(p)
# We make a dummy condition, males are in every even region
my_data[,sex:=ifelse(region_id \%\% 2==0,"m","f")]
p <- ethogamPlot(activity,my_data,condition = sex)
print(p)
p <- ethogamPlot(activity,my_data,condition = sex,error_bar="sm")

# p is simply a ggplot object, so we can change things:
print(p + labs(title="MY own title"))
}
\seealso{
\code{\link{oveviewPlot}} To show per-individual patterns
}

