% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time.R
\name{hours}
\alias{hours}
\title{Trivially converts hours to seconds}
\usage{
hours(x)
}
\arguments{
\item{x}{number of hours}
}
\value{
the corresponding number of seconds
}
\description{
Trivially converts hours to seconds
}
\seealso{
\code{\link{days}} \code{\link{mins}}
}
