% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{loadDAMFile}
\alias{loadDAMFile}
\title{Read a text file formated as DAM2 into a single data table.}
\usage{
loadDAMFile(FILE, tz = "")
}
\arguments{
\item{FILE}{the name of the input file.}

\item{tz}{the time zone of the computer saving the file. By default, \code{tz} is taken from the computer running this function}
}
\value{
a data table with an activity (number of beam crosses) variable, a region_id (channel) variable and a posix time stamp.
}
\description{
This function is used to load data from DAM2 devices as a data.table.
}
\examples{
\dontrun{
FILE <- "Monitor53.txt"
out <- loadDAMFile(FILE)
#histogram of x marginal distribution
hist(out[roi_id == 1, x], nclass=100)
}
\dontrun{
# More realistec example where we have experiemental conditions, and
we want to resample data at 1.0Hz.
# First, the conditions:
conditions <- cbind(roi_id=1:32, expand.grid(treatment=c(T,F), genotype=LETTERS[1:4]))
print(conditions)
}
}
\seealso{
\code{\link{loadMetaData}} To display global informations about the experiment.
}

