% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{loadDAMFile}
\alias{loadDAMFile}
\title{Read a text file formatted as DAM2 into a single data table.}
\usage{
loadDAMFile(FILE, start_date = -Inf, stop_date = +Inf, tz = "",
  verbose = TRUE)
}
\arguments{
\item{FILE}{the name of the input file.}

\item{start_date}{the starting date formated as "yyyy-mm-dd" or "yyyy-mm-dd_hh-mm-ss"}

\item{stop_date}{the last day of the experiment. Same format as \code{start_date}}

\item{tz}{the time zone of the computer saving the file. By default, \code{tz} is taken from the computer running this function}

\item{verbose}{whether to print progress (a logical).}
}
\value{
a data table with an activity (number of beam crosses) variable, a region_id (channel) variable and a posix time stamp.
}
\description{
This function is used to load data from DAM2 devices as a data.table.
}
\examples{
\dontrun{
FILE <- "Monitor53.txt"
out <- loadDAMFile(FILE)
#histogram of x marginal distribution
hist(out[roi_id == 1, x], nclass=100)
}
\dontrun{
# More realistic example where we have experimental conditions, and
we want to resample data at 1.0Hz.
# First, the conditions:
conditions <- cbind(roi_id=1:32, expand.grid(treatment=c(T,F), genotype=LETTERS[1:4]))
print(conditions)
}
}
\seealso{
\code{\link{loadMetaData}} To display global informations about the experiment.
}

