% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ethoscope-data.R
\name{loadEthoscopeData}
\alias{loadEthoscopeData}
\title{Read data from an ethoscope result file}
\usage{
loadEthoscopeData(what, min_time = 0, max_time = Inf,
  reference_hour = NULL, verbose = TRUE, columns = NULL, ncores = 1,
  FUN = NULL, ...)
}
\arguments{
\item{what}{Query describing which file(s) to load and, optionally, associated variables/conditions (see details).}

\item{min_time}{Exclude data before min_time (in seconds).
This time is \emph{relative to the start of the experiment}.}

\item{max_time}{Exclude data after max_time (in seconds).
It is also relative to the start of the experiment.}

\item{reference_hour}{Hour, in the day, to use as ZT0 reference.
When unspecified, time will be relative to the start of the experiment.}

\item{verbose}{whether to print progress (a logical).}

\item{columns}{Optionnal vector of columns to be selected from the db file.
Time (t) is always implicitely selected.}

\item{ncores}{Number of cores to use for optionnal parallel processing.}

\item{FUN}{Optional function to transform the data from each \code{region_id} (i.e. a `data.table``)
immediately after is has been loaded.}

\item{...}{Extra arguments to be passed to \code{FUN}}
}
\value{
A data.table where every row is an individual measurement.
That is a position at a unique time (\code{t}) in a
unique region (\code{region_id}), and from a unique result file/experiment (\code{experiment_id}).
The time is expressed in seconds.
Distance units (e.g. xy position, height/width) are expressed as a fraction of the width of the region
they originate from.
}
\description{
This function is used to convert all the information
contained in a result file generated by the ethoscope platform \url{http://gilestrolab.github.io/ethoscope/}
(i.e a .db file) into an \code{R} \code{data.table}.
}
\examples{
\donttest{
# Case 1: load ALL REGIONS from a SINGLE FILE
# we can download a file from a repo:
#TODO
data_file <- "/somewhere/in/my/computer/file.db"
# `validation_data_file` is simply the path to the .db file in your computer
dt <- loadEthoscopeData(validation_data_file)
print(dt)
# Case 2: load ALL REGIONS from MULTIPLE FILES
# we pass all the files we want to load as the `what` argument
dt <- loadEthoscopeData(paths)
# Note the column `experiment_id` in dt. It tells us which file/experiment 
# each measurement originates from.
print(dt)
}

}
\seealso{
\itemize{
\item \link{buildEthoscopeQuery} to generate a query with the \code{path} of the data file
\item Tutorial for this function \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
\item What queries are \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
\item The structure of the resuting data \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
}
}
