% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ethoscope_io.R
\name{loadEthoscopeMetaData}
\alias{loadEthoscopeMetaData}
\title{Retrieves  metadata from a result file.}
\usage{
loadEthoscopeMetaData(FILE)
}
\arguments{
\item{FILE}{the name of the input file.}
}
\value{
A list containing fields for metadata entries
}
\description{
This function is used to obtain metadata -- such as `time and date of the experiment' , `acquisition device', `version of the software' and such--
embedded in a result file generated by the ethoscope platform.
}
\examples{
\dontrun{
FILE <- "result.db"
out <- loadEthoscopeMetaData(FILE)
names(out)
}
}
\seealso{
\code{\link{loadEthoscopeData}} to load raw data.
}

