% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{loadSampleData}
\alias{loadSampleData}
\title{Retrieves sample/example data contained within in this package.}
\usage{
loadSampleData(names = NULL)
}
\arguments{
\item{names}{The name of the samples to be loaded. When \code{names} is \code{NULL}, the function returns the list of all available samples.}
}
\description{
This function is only for testing (and trying) purposes.  It provides a way to access raw data  (e.g. db files) contained within this package.
}
\seealso{
\code{\link{loadPsvData}} to obtain raw experimental data.
}

