% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/io.R
\name{loadSampleData}
\alias{loadSampleData}
\title{Retreive sample/example data contained in this package.}
\usage{
loadSampleData(names = NULL)
}
\arguments{
\item{names}{The name of the samples to be loaded. When \code{names} is \code{NULL}, the function returns the list of available samples.}
}
\description{
This function is intended to be used in order to try and test this package with a reproducible set of raw db data.
}
\seealso{
\code{\link{loadPsvData}} to obtain raw experimental data.
}

