% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{makeLDAnnotation}
\alias{makeLDAnnotation}
\title{Put white and black bars under a plot to show Dark and Light phases.}
\usage{
makeLDAnnotation(pl, time_conversion_unit = days, period = hours(24),
  offset = 0, size = 0.02)
}
\arguments{
\item{pl}{A \code{ggplot} object to be annotated.}

\item{time_conversion_unit}{The time conversion function used in \code{pl}.}

\item{period}{The period, in seconds}

\item{offset}{A number, lower than the \code{period} to shift the bar left or right}

\item{size}{The hight of the bar. It is expressed in percent of the graph height}
}
\value{
A \code{ggplot} object that can be plotted directly, or modified.
}
\description{
Put white and black bars under a plot to show Dark and Light phases.
}
\examples{
data(sleep_sexual_dimorphism)
my_data <- sleep_sexual_dimorphism
# Fraction of animal asleep over time:
p <- overviewPlot(asleep,my_data,condition=sex)
p <- makeLDAnnotation(p)
print(p)
p <- ethogramPlot(asleep,my_data,condition=sex,error_bar="sem")
p <- makeLDAnnotation(p)
print(p)
}

