% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{maxVelocityClassifierMasked}
\alias{maxVelocityClassifierMasked}
\title{Motion classifier based on maximum velocity, but with a correction factor for variable FPS.}
\usage{
maxVelocityClassifierMasked(data, velocity_correction_coef = 0.003,
  masking_duration = 6)
}
\arguments{
\item{data}{the data.table containing behavioural features used for movement classification.}

\item{velocity_correction_coef}{an empirical coefficient to correct velocity with respect to variable framerate}

\item{masking_duration}{the number of second when any movement is ignored after an interaction (velocity is set to 0).}
}
\value{
a data table with the columns \code{moving} (logical, TRUE iff. motion was detected)
and \code{t_round} (the `rounded' time). 
There is one row per rounded time point.
Also we return the whether an animal has crossed a virtual midline, at x=0.5 (\code{beam_crosses}).
}
\description{
Defines whether an animal is moving according to its subpixel velocity. 
It requires a variable named \code{xy_dist_log10x1000} in the .db file.
}
\seealso{
\code{\link{sleepAnnotation}} to apply this function to all subsequent time windows,
\code{\link{maxVelocityClassifierLegacy}}, the uncorected/unmasked version.
}

