% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time.R
\name{mins}
\alias{mins}
\title{Trivially converts minutes to seconds}
\usage{
mins(x)
}
\arguments{
\item{x}{number of minutes}
}
\value{
the corresponding number of seconds
}
\description{
Trivially converts minutes to seconds
}
\seealso{
\code{\link{days}} \code{\link{hours}}
}

