% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{overviewPlot}
\alias{overviewPlot}
\title{Displays, per individual, the temporal average of a variable of interest.}
\usage{
overviewPlot(y, data, condition = NULL, summary_time_window = mins(30),
  normalise_var_per_id = TRUE)
}
\arguments{
\item{y}{The variable of interest}

\item{data}{The data.table containing the data. It must have a column with the same name as y.}

\item{condition}{An optionnal grouping factor to order rows.}

\item{summary_time_window}{the width (in seconds) of the time window used to draw each pixel.}

\item{normalise_var_per_id}{whether each row is to be normalised, using \code{new_y = (y - mean(y))/sd(y)}.}
}
\value{
A \code{ggplot} object that can be plotted directly or modified.
}
\description{
This function produces a tiled representation in which every row represents one individual (i.e. from a unique combination of region and experiement).
The x axis represents time in days.
The values of the variable of interest are represented by different colour intensity.
}
\examples{
# Load sample data
data(monitor_validation)
my_data <- monitor_validation[,sleepAnnotation(.SD),by=key(monitor_validation)]
# let us have a look of the max velocity as a measure of activity
p <- overviewPlot(max_velocity,my_data)
print(p)
# Let us make a dummy treatment variable.
# For instance, region > 10 have a treatment named ``Drug_1''
my_data[,treatment:=ifelse(region_id > 10,"Control","Drug_1")]
p <- overviewPlot(max_velocity,my_data,condition = treatment)
# See how treatment levels are grouped together:
print(p)
# p is simply a ggplot object, so we can change things:
print(p + labs(title="MY own title"))
}
\seealso{
\code{\link{ethogramPlot}} To show trend by aggregating individuals over time.
}

