% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep-annotation.R
\name{sleepAnnotation}
\alias{sleepAnnotation}
\alias{sleepDAMAnnotation}
\title{Infer sleep from immobility bouts}
\usage{
sleepAnnotation(data, time_window_length = 10, min_time_immobile = mins(5),
  motion_classifier_FUN = maxVelocityClassifierMasked, ...)

sleepDAMAnnotation(data, time_window_length = 60,
  min_time_immobile = mins(5))
}
\arguments{
\item{data}{\code{data.table} representing a \emph{single} animal.
It must contain, at least, the column \code{t} in addition to the variables needed for motion classification.}

\item{time_window_length}{Number of seconds to be used by the motion classifier.
This corresponds to the sampling period of the output data.}

\item{min_time_immobile}{Minimal duration (in s) of a sleep bout.
If longer, an immobility bout is considered as sleep.}

\item{motion_classifier_FUN}{Function used to classify movement.}

\item{...}{Extra arguments to be passed to \code{motion_classifier_FUN}.}
}
\value{
A \code{data.table} similar to \code{data} with additional variables/annotations (i.e. \code{moving} and \code{asleep}).
The resulting data will only have one data point every \code{time_window_length} seconds.
}
\description{
This function uses a motion classifier to first decide whether an animal is moving during a given time window.
Then, it defines sleep as contiguous immobility for a minimal duration.
}
\details{
\code{sleepAnnotation} is typically used for ethoscope data, whilst \code{sleepDAMAnnotation} only works on DAM data.
These function is \emph{rarely used directly}, but rather passed as an argument to a data loading function,
so that analysis can be performed on the go.
}
\examples{
# We stat by making toy data for one animal:
dt_one_animal <- toyEthoscopeData(seed=2)
####### Ethoscope, corrected velocity classification #########
sleep_dt <-  sleepAnnotation(dt_one_animal, masking_duration=0)
print(sleep_dt)
# We make a sleep `barecode'
ggplot(sleep_dt, aes(t,y="Animal 1",fill=asleep)) + 
                                   geom_tile() + scale_x_time()
####### Ethoscope, virutal beam cross classification #########
sleep_dt2 <-  sleepAnnotation(dt_one_animal, 
                             motion_classifier_FUN=virtualBeamCrossClassif)
ggplot(sleep_dt, aes(t,y="Animal 1",fill=asleep)) + 
                                   geom_tile() + scale_x_time()
#' ####### DAM data, de facto beam cross classification ######
dt_one_animal <- toyDAMData(seed=7)
sleep_dt <- sleepDAMAnnotation(dt_one_animal)
ggplot(sleep_dt, aes(t,y="Animal 1",fill=asleep)) + 
                                   geom_tile() + scale_x_time()
}
\seealso{
\itemize{
\item Tutorial for sleep analysis with ethoscopes \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
\item \link{loadEthoscopeData} and \link{loadDAM2Data} to load data and optionally apply these analysis on the fly.
\item \link{maxVelocityClassifierMasked} the motion classifiers that can be used.
}
}
