% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleepAnnotation}
\alias{sleepAnnotation}
\title{Determines whether an animal is asleep}
\usage{
sleepAnnotation(data, time_window_length = 10, min_time_immobile = 60 * 5,
  motion_classifier_FUN = maxVelocityClassifierMasked, ...)
}
\arguments{
\item{data}{the data (i.e a data.table) from a \emph{single} region. It must contain, at least,
the columns `t`, `x` and `y`.}

\item{time_window_length}{The number of seconds to be used by the motion classifier. This corresponds to the sampling period of the output data.}

\item{min_time_immobile}{the minimal duration (in s) after which an immobile animal is scored as `asleep'.}

\item{motion_classifier_FUN}{the function used to classify movement.}

\item{...}{extra arguments to be passed to \code{motion_classifier_FUN}}
}
\value{
A data table similar to \code{data} with additional variables/annotations (i.e. `moving', `asleep').
}
\description{
This function uses a motion classifier to first decide whether an animal is moving during a given time window.
Then, it defines sleep as contiguous immobility for a minimal duration.
}
\note{
The resulting data will only have one data point every \code{time_window_length} seconds.
}
\examples{
# Let us load some sample data
data(tube_monitor_validation)
# We will start only with region 2:
dt_region2 <- tube_monitor_validation[region_id==2,]
sleep_dt <-  sleepAnnotation(dt_region2, masking_duration=0)
print(sleep_dt)
# We make a sleep `barecode'
ggplot(sleep_dt, aes(t,region_id,fill=asleep)) + geom_tile()
# A bit of data.table wizardry to apply that to each experiement and region:
sleep_dt <-  tube_monitor_validation[,sleepAnnotation(.SD, masking_duration=0),by=key(tube_monitor_validation)]
# The same bare code for all regions
ggplot(sleep_dt, aes(t,region_id,fill=asleep)) + geom_tile()
}
\seealso{
\code{\link{loadEthoscopeData}} to load data and optionally apply analysis on the fly.
}

