% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-data.R
\name{toyActivityData}
\alias{toyActivityData}
\title{Generate toy activity and sleep data mimiking Drosophila behaviour in tubes}
\usage{
toyActivityData(query = NULL, seed = 1, rate_range = 1/c(60, 10),
  duration = days(5), sampling_period = 10, ...)
}
\arguments{
\item{query}{Query (i.e. a \code{data.table} where every row defines an animal).
Typically queries have, at least, the columns \code{experiment_id} and \code{region_id}.
The default value (\code{NULL}), will generate data for a single animal.}

\item{seed}{Random seed used.}

\item{rate_range}{a parameter defining the boundaries of rate at which animals wake up.
It will be uniformely distributed between animals, but fixed for each animal.}

\item{duration}{Length (in second) of the data to generate.}

\item{sampling_period}{sampling period (in second) of the resulting data.}

\item{...}{Additional arguments to be passed to \code{simulateAnimalActivity}}
}
\value{
A behavioural \code{data.table} with the query columns as key and two activity variables (\code{moving} and \code{asleep}).
}
\description{
This function generates random data that emulates some of the features of fruit fly activity and sleep.
This is designed \strong{exclusively to provide material for examples and tests} since it generates  "realistic" datasets of arbitrary length.
}
\examples{
query<- data.table(experiment_id="toy_experiment",
                   region_id=1:20, 
                   condition=c("A","B"))
print(query)
dt <- toyActivityData(query,3)
print(dt)
}
\author{
Quentin Geissmann (\email{qgeissmann@gmail.com})
}
